requirements.txt
absl-py==2.0.0
adal==1.2.7
addict==2.4.0
anykeystore==0.2
apex==0.1
argcomplete==3.5.0
argparse==1.4.0
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.30.2
azure-graphrbac==0.61.1
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==26.0.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azureml==0.2.7
azureml-core==1.57.0
backports-tempfile==1.0
backports-weakref==1.0.post1
batchgenerators==0.25
bcrypt==4.2.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
cmake==3.27.5
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.2.1
cryptacular==1.6.2
cryptography==43.0.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docker==7.1.0
easydict==1.11
efficientnet-pytorch==0.7.1
einops==0.7.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.12.4
fonttools==4.51.0
fsspec==2024.3.0
future==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
greenlet==3.0.3
grpcio==1.59.3
h5py==3.10.0
huggingface-hub==0.21.4
humanfriendly==10.0
hupper==1.12.1
idna==3.4
imageio==2.33.0
imbalanced-learn==0.12.3
imblearn==0.0
importlib-metadata==6.8.0
importlib-resources==6.4.0
ipdb==0.13.13
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
jeepney==0.8.0
jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.3.0
kiwisolver==1.4.5
knack==0.11.0
lazy-loader==0.3
lightning-utilities==0.11.7
linecache2==1.0.0
lit==16.0.6
markdown==3.5.1
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
medpy==0.5.2
mmcv==2.2.0
mmengine==0.10.4
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
munch==4.0.0
ndg-httpsclient==0.5.1
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.2.2
paramiko==3.4.1
parso==0.8.4
pastedeploy==3.1.0
pathspec==0.12.1
pbkdf2==1.3
pexpect==4.9.0
pillow==10.0.1
pkginfo==1.11.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.2
portalocker==2.10.1
pretrainedmodels==0.7.4
prettytable==3.10.0
prompt-toolkit==3.0.48
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pygments==2.18.0
pyjwt==2.9.0
pynacl==1.5.0
pyopenssl==24.2.1
pyparsing==3.1.2
pyramid==2.0.2
pyramid-mailer==0.15.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pyyaml==6.0.1
repoze-sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
secretstorage==3.3.3
segmentation-models-pytorch==0.3.3
simpleitk==2.4.0
six==1.16.0
sqlalchemy==2.0.29
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
tomli==2.0.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchmetrics==1.4.2
torchvision==0.15.2+cu117
tqdm==4.66.2
traceback2==1.4.0
traitlets==5.14.3
transaction==4.0
translationstring==1.4
triton==2.0.0
typing-extensions==4.8.0
tzdata==2024.1
unittest2==1.1.0
urllib3==2.0.4
velruse==1.1.1
venusian==3.1.0
wcwidth==0.2.13
webob==1.8.7
werkzeug==3.0.1
wtforms==3.1.2
wtforms-recaptcha==0.3.2
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0
zope-deprecation==5.0
zope-interface==6.2
zope-sqlalchemy==3.1